fix(upload.sh): avoid running nodejs when there is nothing inside prepare

+++ upload.sh
@@ -2,11 +2,13 @@
 
 set -eu -o pipefail
 
-touch ./store/cids
-node index.js ./prepare/ > store.log
-cat store.log >> ./store/cids
+if [ -d "prepare" ] && (find prepare -type f | grep '' >/dev/null); then
+  touch ./store/cids
+  node index.js ./prepare/ > store.log
+  cat store.log >> ./store/cids
 
-for file in ./prepare/* ; do
-  rm "$file"
-done
+  for file in ./prepare/* ; do
+    rm "$file"
+  done
+fi
